Add files via upload
Claro! Aqui está o texto completo traduzido:

"In this code, jump height was implemented by measuring the maximum height reached by the center of mass. To collect the video, the individual should remain between 3 to 4 meters from the camera, stay still for 1 second, and perform a countermovement jump. Better accuracy in the results is achieved with video acquisition frequencies of 120 Hz and 240 Hz. It is necessary to have the individual's weight and height for scaling the center of mass."